+++ tests/system_tests/test_sweep/test_wandb_agent.py
@@ -194,81 +194,12 @@ def _write_ignore_signals_child_script(path: Path) -> Path:
     return script
 
 
-def _write_term_timeout_parent_script(path: Path) -> Path:
-    """Similar to _write_signal_parent_script, but uses an Agent to manage the
-    AgentProcess lifecycle, isntead of using AgentProcess directly, which allows us to
-    pass in a term_timeout parameter."""
-
-    script = path / "parent_term_timeout.py"
-    script.write_text(
-        textwrap.dedent(
-            """
-            import os
-            import queue
-            import signal
-            import sys
-            import threading
-            import time
-
-            from wandb.wandb_agent import Agent
-
-            child_script = sys.argv[1]
-            term_timeout = int(sys.argv[2])
-
-            class _StubApi:
-                def sweep(self, sweep_id, spec):
-                    return None
-
-                def register_agent(self, host, sweep_id=None):
-                    return {"id": "agent-1"}
-
-                def agent_heartbeat(self, agent_id, spec, run_status):
-                    return []
-
-            # Seed agent command queue to run the child script
-            command_queue = queue.Queue()
-            command_queue.put(
-                {
-                    "type": "run",
-                    "run_id": "run-1",
-                    "program": child_script,
-                    "args": {},
-                    "resp_queue": queue.Queue(),
-                }
-            )
-
-            agent = Agent(
-                _StubApi(),
-                command_queue,
-                sweep_id="sweep-1",
-                term_timeout=term_timeout,
-                forward_signals=True,
-            )
-
-            # Send a SIGINT after 1 second. This is eaten by then child process but
-            # puts the Agent into the tier 1 waiting state.
-            threading.Timer(1.0, lambda: os.kill(os.getpid(), signal.SIGINT)).start()
-
-            start = time.monotonic()
-            agent.run()
-            elapsed = time.monotonic() - start
-
-            procs = list(agent._run_processes.values())
-            child = procs[0]
-
-            returncode = child.wait(timeout=5)
-            if returncode != -signal.SIGKILL:
-                sys.exit(2)
-
-            # We expect that the child terminated roughly around the term_timeout time
-            if elapsed > term_timeout * 2:
-                sys.exit(3)
-            if elapsed < term_timeout * 0.5:
-                sys.exit(4)
-            """
-        ).strip()
-    )
-    return script
+# Parent script that used in the term_timeout end-to-end test. Unlike
+# _write_signal_parent_script, it uses an Agent to manage the AgentProcess
+# lifecycle (rather than AgentProcess directly) so it can pass a term_timeout.
+_TERM_TIMEOUT_PARENT_SCRIPT = (
+    Path(__file__).parent / "mock_scripts" / "parent_script_with_term_timeout.py"
+)
 
 
 @pytest.mark.skipif(
@@ -307,8 +238,8 @@ def test_agent_term_timeout_escalates_to_sigkill(tmp_path):
     escalated straight to SIGKILL rather than re-sent a polite SIGTERM."""
 
     child_script = _write_ignore_signals_child_script(tmp_path)
-    parent_script = _write_term_timeout_parent_script(tmp_path)
-    term_timeout = 3
+    parent_script = _TERM_TIMEOUT_PARENT_SCRIPT
+    term_timeout = 1
 
     repo_root = Path(__file__).resolve().parents[3]
     env = os.environ.copy()
